darkguard-backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
supabase==2.11.0
groq==0.15.0
httpx==0.28.1
redis==5.2.1
celery[redis]==5.4.0
weasyprint==63.1
Jinja2==3.1.5
scikit-learn==1.6.1
skl2onnx==1.18.0
onnxruntime==1.20.1
python-multipart==0.0.20
playwright==1.49.1
Pillow==11.1.0
